Beginning to interface with Zeta
I have fixed a number of bugs. Most notably, we can construct all the
subcharts of a locally monomial chart. We can construct their
integrands, and we can construct placeholders for the counts of
p-rational points on varieties. The next thing I want to do is put a slightly smarter p-rational point counter. Of course, in general it will be a placeholder or it will ask the user, but there are many common varieties we consider.