A better p-rational counter
It's not far better, but it is a little smarter. It can handle a little
more than linear systems. What is needed now is a lookup system: if the
user provides a count of the p-rational points, then we should take
advantage of it. In addition, the polynomials pop up in more than one
place, so this will speed up (human) computation time.