p-rational point database
Added the infrastructure for a database to store the p-rational points
of a variety. At the moment, it does not actually check the database. It
will do a superficial check, and if it passes this, then it just returns
that it did not find anything. Eventually, we will run through a
suitable permation group. This should greatly decrease the amount of
human involvement.