Cleaned up intersection lattices
The intersection lattice from Singular comes with lots of redundancies,
and I believe we have removed nearly all redundancies. This should
improve the timing significantly, but more importantly, this should help
with bug testing as there are fewer pieces to manage. Next up is dealing
with the discrepancy from the p-rational point counts.