Property Test: IsLocalSubringZF_Zn
Added a property test for symbolic expressions in at most 2 variables to check if it is a local (unital) subring zeta function of Z^n, for a given n. For n < 5, it compares against what is known. For n at least 5, it compares the first few terms with what is expected.